New commands, fixes, tweaks and the beginnings of support for a new botman mod!
Added custom folder to the github repo containing custom_fuctions.lua, customIRC.lua, and gmsg_custom.lua.  These have been around for a long time and are places where you can add your own custom code to the bot without risk of a future bot update deleting your code.  The update only replaces code in the scripts folder.  The custom folder lives above scripts.

New commands:

/remove entity {optional entity id}
The bot will despawn any trader within 2 blocks of you in-game or an entity by id if given.
This command makes it a little easier to de-spawn traders.  Just stand next to them and use this command without an id.  You will need an id if you have other traders right next to you.

/set bot name colour {hex colour code}
Set the colour of the bot's name.  Requires the botman mod.  With the botman mod installed, you can change messages from Server: to use the name of your bot.  This command allows you to set a colour for the bot's name in chat.  Note that the mod also supports a colour for the bot's name in private messages.  Currently I have only coded support for public messages.

/set max log days {number}
The default is 14 days. Setting this too long will result in lots of log files on disk.  To prevent this causing issues on my hosted bots, the max you can set this is 60 days.  To set it higher you need to edit telnetlogkeepdays in the server table in the database.

New command to set a custom server vote reward.  Note that random and list options are not coded yet but the rest is finished.

/set server vote reward {default/random/list/item/entity}
/set server vote reward random quantity {number} quality {number}
The default reward is the sc_General supply crate.
If you set a random reward you can optionally set a quantity and quality.
Quality is random if not specified.
Quantity is the number of random items, not a quantity of the same item (except by chance). If quantity is not set the bot will give between 3 and 5 random items.
If you set the reward as list you will need additional commands to manage the reward items. For commands type {#}help reward list.
If you set the reward as item, the player will be given the item that you specify.
If you have a custom entity (eg sc_General2) type {#}set server vote reward entity sc_General2.

/test vote reward
In-game only and restricted to server owners.  This command allows you to test spawning the server vote reward.  Note that it currently also flags you as having actually voted.  I will fix that in a future update.

/enable/disable fetch
Allow or block players using the fetch command to teleport friends to them.  This command allows you to still have player to player teleporting enabled while disabling /fetch.

Added support for the botman mod's console commands bm-say and bm-sayprivate which allows Server: to be replaced with Botman:

Fixed bug in the inventory scanner.  I am also doing a major re-code of the scanner but I have held that back until it is finished.

Fixed an issue with reading GG in API mode.  The BC mod JSON encodes the result which is fine until the server join message includes \r\n's then the parser breaks and the rest of gg can't be read.  I fixed this issue by forcing the bot to use telnet to read gg.

I have added a lua file for botman mod chat commands called gmsg_botman.lua.  Currently there are no commands in it.

The chat parser has some code changes to support messages coming from Server: or your bot's name.  On IRC they still show up as coming from Server: but in-game they are altered.

Bot command support for coppi's mod and djkrose's scripting mod has been disabled as those are only relevant in A16.  Coppi's mod has been rolled into the botman mod and commands will re-appear there.

I fixed some bugs in /claim vote when using a custom entity as the vote reward.  I still have code to add to complete the new vote reward options.

Added a minimum value check on /set max uptime so you can't set it to 0.

The IRC command 'near player joe' now lists entities near them.  Note that each time you use this command it does a scan for entities but you won't see the result of the scan unless you repeat the command.  If you haven't used the command in a while and it does list entities, the list will be out of date.

Added several new fields to the server table, some of which have no code to support them yet.

Added some web interface support for the botman web interfaces (2) that are coming.

Various small bug fixes and tweaks.