Bug fixes, more panel support, new commands
First off we have lost a friend at the beginning of 2020.  KenJustKen has passed away.  I hid a note in the code near a bot feature that was important to him.  I know this because he bugged me about it a lot until I relented and made some changes. xD
Wherever you are now Ken, I am sure that you are still watching :O

The admin command '/bases' now shows base protection.

Fixed a bug that prevented admins setting a base for another player with /setbase {player}

Tweaked how the bot sets up and starts using Alloc's web API.  It should be a lot better at that now.

Added help for the dig command.  You've always been able to swap a block for another block but the help didn't mention this.

Added clan command support..
/set max clans {number of clans} - Sets the maximum number of clans allowed.
/set max clan players {number of players} - Set the maximum number of players a clan can have.
/set min clan level {player level} - Set the minimum player level required to join a clan.
/enable/disable clans - Enable or disable the clan feature in the Botman mod.

Added support for some of the reset regions commands in the Botman mod (this is a work in progress) - more updates to follow.
/enable/disable reset regions - Enable or disable the Botman mod's reset regions feature.
This is very similar to the bot's reset zones except that they are limited to whole regions and can actually reset the region or just prefabs or just remove claims. (requires the harmony.dll which must be installed in the folder 7Daystodieserver_data/managed).
/reset regions remove/leave claims (default is leave claims) - Reset regions can remove or leave claim blocks.

Some fixes to the bot's arena game.  It should work better though I've had some new reports so we'll see.

Tweaked the chat parser to fix some issues that mostly turned out to be mod issues. That drove me insane for quite a while :O
Also the ability to teleport to a friend by typing their name eg /joe has been replaced with /visit joe (or alternatively /goto joe).  This was done to fix issues with player names occasionally clashing with commands in other managers/mods and occasionally clashes in the bot.
Note that if a player tries to teleport to a friend the old way, the bot will tell them unknown command, then it will tell them the new way to teleport to friends.

The idle kicker's timer can now be set and also it can be told to kick any time, not just when the server is full.

Fixed a command clash between /spawn and /spawn horde.  There were some other tweaks to lobby and spawn but I am not sure if those are finished or not.

The command /invite {player} has changed to /irc invite {player}

When clearing all reset zones, the bot now tells the botman mod to forget them all as well.  Ooops!

Adding and removing reset zones also adds them via the botman mod (if installed).

Added /set idle kick {seconds} (default 900. 15 minutes) - How many seconds a player can be idle for before being kicked from the server. Does not include joining players that have not spawned yet.

Added /allow/disallow idling (allowed is default) - If idle kick is enabled the default is to only kick idle players when the server is full. By setting disallow idling, players can be kicked for idling any time.

Added /visit map command which has several variations..
/visit map
/visit map x1 z1 x2 z2
/visit map range {distance}
/visit map stop
Make the server explore the map while you hit up some zombie chicks.
If you add the optional word 'check' at the end of the command it will also check chunk density.
To visit the entire map do not include coordinates or a range.

The admin command '/tp' can now teleport admins to the middle of a region.
/tp region {region X} {region Z}

The cooldown for waypoints is halved for donors.

More bot support for the web panel has been added with more to come.  Soon the panel will be able to show a list of who is playing right now on the server.

Fixed a bug with restoring a specific player from a bot backup.

Major overhaul of the bot's database maintenance code to fix an issue with newer versions of MariaDB and MySQL being more strict about SQL stuff.  The bot should work in old and new releases of the database engines equally well.

Fixed a bug with saving a player record.  Kinda important that that works. It was tripping up on a new field type that I added (date).

Fixed a bug causing the shop to not index shop items giving them all the same item code.

Tweaks to the bot's startup code for debugging and the api.

The /return cooldown is half for donors.

The bot doesn't send the mem command anymore.  It can crash servers.

Attempted to fix a bug that randomly causes the bot to think players are new again when they are not and it resets some stuff.  This is caused by what I think is a Pimps bug.  Occasionally the Inf player joined line has a rogue new line character in it which causes the line to split.  I say attempted because I am still getting reports that this is happening.  Welp back to pub er.. drawing board.

In IRC the bot doesn't report disconnecting players that didn't join (such as banned players). So you can no longer watch them attempting to join over and over fruitlessly.

The /pack command cooldown is halved for donors.