Bug fixes mostly
Fixed /stuck command so it works immediately and also only sends the player up a few blocks.

Added command /enable pvp visit.  Allows the /visit command in pvp rules. Normally visit is not allowed.

Fixed oops in function loadGimmePrizes.  It was mixed up with the shop table.  I had to allow for the stuff up when reloading the shop from backups too.  Fun times.

The nightly data purge that the bot does on old tracking data now includes the performance and events tables which are both rigged to keep only the last 90 days of data.

Fixed an issue with the persistent queue used by the bot for future dated commands so that commands wouldn't block other commands that were added later but were timed to run earlier.

Some other minor tweaks to the daily logging so that write permissions on the log folder or the complete absence of the log folder doesn't cause problems.