Fixes
Fixed map permissions which were too high.

Fixed typo in inventory scanner that prevented banning for certain items. It was sending players to timeout instead.

Added auto fix for when the IRC alerts channel and watch channel get accidentally swapped (happened in panel).

Added terrainFiller and terrainRemove to the bad items list set to ban if found in player inventory.  It is a favourite tool of hackers lately.

The bot now reads the anticheat report every minute as a fallback method to catch hackers if live reporting fails for some reason.

Bans were not being loaded into the lua table bans after reading bans from the server which caused disagreement between the bans table and the in-memory lua table.

Improved handling of the anticheat report when the bot reads it.  It wasn't catching them all.  Do not invite the bot to play Pokemon.