Bugfixes and changes
Bug fixes

Releasing a prisoner now zeroes their bail

Fixed an issue with the bot reporting to players crazy long wait times for cooldown timers when the real wait was less than an hour.

Fixed an issue where a line 1 block wide could exist between 2 parts of the map where on one side the rules were pve and on the other they were pvp.  Now North and East covers that extra block.

Fixed a bug in gimme where the randomly chosen zombie's entity id was wrong.

Fixed an issue where the panel would update a table and tell the bot to reload it but when the bot did that, the database was still returning the old data.  Adding a 3 second delay allows the database to finish updating before the bot does the reload.

Fixed an issue where the bot would detect an anticheat entry and act on it a lot which caused tons of unnecessary spam.  Now it maintains a tempory table so that it only triggers if the steam id from the anticheat event hasn't been recorded yet.

Changes

/resettimers now resets all cooldowns for a player.  Previously it just reset base and gimme cooldowns.

Added shop to the options of things that can be restored from the bot's backups.

Replaced the spawnableItems table with a Lua table which is much faster since it only exists in memory.  It used to be a MySQL memory table but it caused significant delays of more than a minute every time the bot was asked to read the list if items known to the server.  Now that tasks takes about 1 second. As a result and to fix shop issues, I am now validating the shop items much more regularly which should ensure that the shop never contains items that can't be spawned.

When donors expire (the status not the people) the bot no longer wipes their waypoints.  Instead they won't be able to create new waypoints until they have deleted enough that they have less waypoints than their new non-donor max waypoints.

A19 introduced a small change to the output of admin list which meant the a18 code couldn't read it.  Now both a18 and a19 code branches can handle the old format and the new format, however if the bot detects the server running A19 and the bot is on the a18 code branch it will automatically switch to the a19 code branch to ensure that it doesn't any future A19 only code changes.

The bot reads the Botman mod's config file.  Now the config file's setting will replace some settings in the bot such as the bot's name, chat hidden state, command prefix and the state of the anticheat system.

Attempted bug fixes

Occasionally a player can be reset by the bot as if they were new which causes a lot of gnashing of teeth and stuff.  The bot now counts the number of times the steam id appears in the events table and if there are more than 3 it will not initialise the player as new.  Hopefully that fixes that issue.